refactor: domain에 있던 출력 책임을 view 패키지로 이동
