+++ src/main/java/view/PieceAppearance.java
@@ -1,6 +1,7 @@
-package domain.piece;
+package view;
 
 import domain.game.Team;
+import domain.piece.PieceDefinition;
 
 public interface PieceAppearance {
     String colorize(Team team, PieceDefinition type);
